+++ jpro-auth/example/src/main/java/one/jpro/platform/auth/example/login/GoogleLoginApp.java
@@ -8,7 +8,7 @@ import one.jpro.platform.auth.example.login.page.ErrorPage;
 import one.jpro.platform.auth.example.login.page.LoginPage;
 import one.jpro.platform.auth.example.login.page.SignedInPage;
 import one.jpro.platform.auth.example.oauth.OAuthApp;
-import one.jpro.platform.auth.routing.AuthBasicOAuth2Filter;
+import one.jpro.platform.auth.routing.AuthBasicOAuth2Transformer;
 import one.jpro.platform.auth.routing.AuthUIProviders;
 import one.jpro.platform.auth.routing.UserSession;
 import one.jpro.platform.routing.Response;
@@ -75,7 +75,7 @@ public class GoogleLoginApp extends RouteApp {
                 .and(Route.get("/", request -> Response.node(new LoginPage(googleAuthProvider, uiProvider))))
                 .when(request -> isUserAuthenticated(), Route.empty()
                         .and(Route.get("/user/signed-in", request -> Response.node(new SignedInPage(this, googleAuthProvider)))))
-                .transform(AuthBasicOAuth2Filter.create(googleAuthProvider, userSession,
+                .transform(AuthBasicOAuth2Transformer.create(googleAuthProvider, userSession,
                         user -> Response.redirect("/user/signed-in"),
                         error -> Response.node(new ErrorPage(error))))
                 .transform(DevTransformer.create());